Use package extensions
