+++ Project.toml
@@ -1,37 +1,43 @@
 name = "ExpressionTreeForge"
 uuid = "93090adf-0e31-445f-8c8f-44d91f61d7ad"
-authors = ["raynaudp "]
+authors = ["raynaudp"]
 version = "0.3.0"
 
 [deps]
-ADNLPModels = "54578032-b7ea-4c30-94aa-7cbd1cce6c9a"
 ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210"
-JuMP = "4076af6c-e467-56ae-b986-b466b2749572"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 MathOptInterface = "b8f27783-ece8-5eb3-8dc8-9495eed66fee"
-NLPModelsJuMP = "792afdf1-32c1-5681-94e0-d7bf7a5df49e"
 ReverseDiff = "37e2e3b7-166d-5795-8a7a-e32c996b4267"
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
+
+[weakdeps]
+ADNLPModels = "54578032-b7ea-4c30-94aa-7cbd1cce6c9a"
+JuMP = "4076af6c-e467-56ae-b986-b466b2749572"
+NLPModelsJuMP = "792afdf1-32c1-5681-94e0-d7bf7a5df49e"
 Symbolics = "0c5d862f-8b57-4792-8d23-62f2024744c7"
+
+[extensions]
+ExpressionTreeForgeJuMPExt = "JuMP"
+ExpressionTreeForgeADNLPModelsExt = ["ADNLPModels", "Symbolics"]
+ExpressionTreeForgeNLPModelsJuMPExt = "NLPModelsJuMP"
 
 [compat]
 ADNLPModels = "0.8"
 ForwardDiff = "0.10,1"
+LinearAlgebra = "1.10"
 JuMP = "1"
 MathOptInterface = "1"
 NLPModelsJuMP = "0.13"
+OptimizationProblems = "0.9"
 ReverseDiff = "1.14"
-Symbolics = "5,6"
-julia = "^1.10.0"
+SparseArrays = "1.10"
+Symbolics = "6"
+Test = "1.10"
+julia = "1.10"
 
 [extras]
-ADNLPModels = "54578032-b7ea-4c30-94aa-7cbd1cce6c9a"
-MathOptInterface = "b8f27783-ece8-5eb3-8dc8-9495eed66fee"
-Symbolics = "0c5d862f-8b57-4792-8d23-62f2024744c7"
-NLPModelsJuMP = "792afdf1-32c1-5681-94e0-d7bf7a5df49e"
 OptimizationProblems = "5049e819-d29b-5fba-b941-0eee7e64c1c6"
-SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["Test", "SparseArrays", "MathOptInterface", "Symbolics", "ADNLPModels", "OptimizationProblems", "NLPModelsJuMP"]
+test = ["ADNLPModels", "OptimizationProblems", "NLPModelsJuMP", "JuMP", "Symbolics", "Test"]
