+++ src/ExpressionTreeForge.jl
@@ -1,5 +1,6 @@
 module ExpressionTreeForge
 
+include("get_expression_tree.jl")
 include("type_expr/ordered_include.jl")
 include("node_expr_tree/ordered_include.jl")
 include("tree/ordered_include.jl")
@@ -11,8 +12,6 @@ using .M_implementation_complete_expr_tree
 using .M_bound_propagations, .M_convexity_detection
 
 include("export.jl")
-
-include("NLPSupport/ordered_include.jl")
 
 export create_bounds_tree, get_bounds, set_bounds!
 export Type_node, Complete_expr_tree, Pre_compiled_tree, Pre_n_compiled_tree, Type_calculus_tree
